Add custom errors to build class
This commit adds our error classes when raising an error in the build
class. This is important for specs. Before the specs expected any
errors and were therefore passing. Upon putting in custom errors I
found a bunch of errors in the code that were passing the specs before.
It's important to implement custom errors as fast as possible and
continue to use them, so we don't make more silly errors.